Add note how to keep column order with piping
Losing column order via --pretty as specified in the SELECT is clear no go,
but of course, we want to see tables as tables and not as json.

So, I added an example of how to pipe it into a specialiced tabular
tool (here: visidata, what is a vim like spreadsheet tool) that shows
the table beautiful and efficient in the specified order.

I'm not sure whether this high level tool (with similar functionality as
excel) is really the best example, but wanted to make at least a
suggestion as a starting point for discussion.

As said, losing the order in a pretty view is all, but certainly not
pretty.